requirements.txt
matplotlib==3.3.0
streamlit==0.63.1
Keras==2.3.1
numpy==1.17.3
Pillow==7.2.0
